fix(search): keyboard support for enter key
